requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
flask>=2.0.0
